🚚 Relocate components into src folder
